+++ internal/provider/cm/resource_scheduler.go
@@ -1,4 +1,4 @@
-package scheduler
+package cm
 
 import (
 	"context"
